[#105] Rename arm_control package to mcu_control
