+++ robot/rospackages/src/mcu_control/msg/RoverGoal.msg
