+++ robot/rospackages/src/mcu_control/scripts/ScienceNode.py
@@ -11,15 +11,15 @@ import serial.tools.list_ports # pyserial
 #import from mcuSerial import McuSerial # this isn't anything yet, just a copy of uart.py
 
 import rospy
-from std_msgs.msg import String, Header, Float32
-from sensor_msgs.msg import JointState
-from arm_control.srv import *
+from std_msgs.msg import String, Header
+# TODO: change this package name to something more generic
+from mcu_control.srv import *
 
 global ser # make global so it can be used in other parts of the code
-mcuName = 'arm'
 
 # 300 ms timeout... could potentially be even less, needs testing
 timeout = 0.3 # to wait for a response from the MCU
+mcuName = 'science'
 
 # todo: test ros+website over network with teensy
 # todo: make a MCU serial class that holds the port initialization stuff and returns a reference?
@@ -37,10 +37,10 @@ def init_serial():
         rospy.loginfo('Using UART by default')
     if len(myargv) > 1:
         if myargv[1] == 'uart':
-            usb=False; uart=True
+            usb = False; uart = True
             rospy.loginfo('Using UART and 115200 baud by default')
         elif myargv[1] == 'usb':
-            usb=True; uart=False
+            usb = True; uart = False
             rospy.loginfo('Using USB and 115200 baud by default')
         else:
             rospy.logerr('Incorrect argument: expecting "usb" or "uart"')
@@ -56,7 +56,7 @@ def init_serial():
             rospy.loginfo("%d USB device(s) detected", len(ports))
             for portObj in ports:
                 port = portObj.name
-                rospy.loginfo('Attempting to connect to /dev/'+port)
+                rospy.loginfo('Attempting to connect to /dev/' + port)
                 ser = serial.Serial('/dev/' + port, baudrate)
 
                 rospy.loginfo("clearing buffer...")
@@ -71,11 +71,11 @@ def init_serial():
                     while (time.time()-startListening < timeout):
                         if ser.in_waiting: # if there is data in the serial buffer
                             response = ser.readline().decode()
-                            rospy.loginfo('response: '+response)
+                            rospy.loginfo('response: ' + response)
                             if mcuName in response:
-                                rospy.loginfo(mcuName+" MCU identified!")
+                                rospy.loginfo(mcuName + " MCU idenified!")
                                 rospy.loginfo('timeout: %f ms', (time.time()-startListening)*1000)
-                                rospy.loginfo('took %f ms to find the '+mcuName+' MCU', (time.time()-startConnecting)*1000)
+                                rospy.loginfo('took %f ms to find the " + mcuName +  " MCU', (time.time()-startConnecting)*1000)
                                 return
         else:
             rospy.logerr("No USB devices recognized, exiting")
@@ -83,7 +83,7 @@ def init_serial():
 
     elif uart:
         port = 'ttySAC0'
-        rospy.loginfo('Attempting to connect to /dev/'+port)
+        rospy.loginfo('Attempting to connect to /dev/' + port)
         try:
             ser = serial.Serial('/dev/' + port, baudrate)
         except:
@@ -101,72 +101,95 @@ def init_serial():
             ser.write(str.encode('who\n'))
             while (time.time()-startListening < timeout):
                 if ser.in_waiting:
-                    dat='';data=None
+                    dat = ''; data = None
                     try:
                         dat = ser.readline().decode()
                         data = stripFeedback(dat)
                     except:
                         rospy.logwarn('trouble reading from serial port')
                     if data is not None:
-                        if mcuName in data:
-                            rospy.loginfo(mcuName+" MCU identified!")
+                        if mcuName in response:
+                            rospy.loginfo(mcuName + " MCU idenified!")
                             rospy.loginfo('timeout: %f ms', (time.time()-startListening)*1000)
-                            rospy.loginfo('took %f ms to find the '+mcuName+' MCU', (time.time()-startConnecting)*1000)
+                            rospy.loginfo('took %f ms to find the ' + mcuName + ' MCU', (time.time()-startConnecting)*1000)
                             return
                     else:
-                        rospy.loginfo('got raw message: '+dat)
+                        rospy.loginfo('got raw message: ' + dat)
 
     rospy.logerr('Incorrect MCU connected, terminating listener')
     sys.exit(0)
 
+# it may cause issues to have responses that are so similar
 requests = {
     'ping' : ['pong'],
-    'who' : ['arm'],
-    'loop open' : ['open'],
-    'loop closed' : ['closed'],
-    'armspeed' : ['Alert', 'Success'],
-    'reboot' : ['rebooting']
+    'who' : ['science'],
+    'activate' : ['activated'],
+    'deactivate' : ['deactivated'],
+    'stop' : ['stopped'],
+    'active' : ['activated1', 'activated0'],
+    'dccw' : ['dccw done'],
+    'dcw' : ['dcw done'],
+    'dd' : ['CW', 'CCW'],
+    'eup' : ['eup done'],
+    'edown' : ['edown done'],
+    'ed' : ['UP', 'DOWN'],
+    'pd0' : ['OUT'],
+    'pd1' : ['IN'],
+    'p1' : ['p1 done'],
+    'p2' : ['p2 done'],
+    'p3' : ['p3 done'],
+    'p4' : ['p4 done'],
+    'p5' : ['p5 done'],
+    'led1' : ['led1 done'],
+    'led1s' : ['led1s done'],
+    'led2' : ['led2 done'],
+    'led2s' : ['led2s done'],
+    'dgo' : ['dgo done'],
+    'ds' : ['ds done'],
+    'drillspeed' : ['drillspeed done']
 }
+
 def handle_client(req):
     global ser # specify that it's global so it can be used properly
     global reqFeedback
     global reqInWaiting
-    armResponse = ArmRequestResponse()
+    scienceResponse = ScienceRequestResponse()
     timeout = 0.1 # 100ms timeout
     reqInWaiting=True
     sinceRequest = time.time()
-    rospy.loginfo('received '+req.msg+' request from GUI, sending to arm Teensy')
-    ser.write(str.encode(req.msg+'\n')) # ping the teensy
-    while armResponse.success is False and (time.time()-sinceRequest < timeout):
+    rospy.loginfo('received ' + req.msg + ' request from GUI, sending to " + mcuName + " MCU')
+    ser.write(str.encode(req.msg + '\n')) # ping the teensy
+    while scienceResponse.success is False and (time.time()-sinceRequest < timeout):
         if reqFeedback is not '':
-            print(reqFeedback)
+            print('reqFeedback', reqFeedback)
             for request in requests:
                 for response in requests[request]:
                     if request == req.msg and response in reqFeedback:
-                        armResponse.response = reqFeedback
-                        armResponse.success = True #a valid request and a valid response from the
+                        scienceResponse.response = reqFeedback
+                        scienceResponse.success = True # a valid request and a valid response from the
                         break
-            if armResponse.success:
+            if scienceResponse.success:
                 break
             else:
-                armResponse.response += reqFeedback
-        rospy.Rate(100).sleep()
-    rospy.loginfo('took '+str(time.time()-sinceRequest)+' seconds, sending this back to GUI: ')
-    rospy.loginfo(armResponse)
+                scienceResponse.response += reqFeedback
+        rospy.Rate(100).sleep() # test with 50? / 50ms timeout above
+    rospy.loginfo('took ' + str(time.time()-sinceRequest) + ' seconds, sending this back to GUI: ')
+    rospy.loginfo(scienceResponse)
     reqFeedback=''
     reqInWaiting=False
-    return armResponse
+    return scienceResponse
 
 def subscriber_callback(message):
     global ser # specify that it's global so it can be used properly
-    rospy.loginfo('received: '+message.data+' command from GUI, sending to arm Teensy')
-    command = str.encode(message.data+'\n')
+    rospy.loginfo('received: ' + message.data + ' command from GUI, sending to ' + mcuName + ' Teensy')
+    command = str.encode(message.data + '\n')
     ser.write(command) # send command to teensy
     return
 
+"""
 def publish_joint_states(message):
     # parse the data received from Teensy
-    lhs,message = message.split('Motor Angles: ')
+    lhs,message = message.split('Science Data: ')
     angles = message.split(', ')
     # create the message to be published
     msg = JointState()
@@ -181,9 +204,10 @@ def publish_joint_states(message):
     anglePub.publish(msg)
     rospy.logdebug(msg.position)
     return
+"""
 
 def stripFeedback(data):
-    startStrip='ARM '
+    startStrip=mcuName.upper() + ' '
     endStrip='\r\n'
     if data.startswith(startStrip) and data.count(startStrip) == 1:
         if data.endswith(endStrip) and data.count(endStrip) == 1:
@@ -193,76 +217,54 @@ def stripFeedback(data):
     return None
 
 if __name__ == '__main__':
-    node_name = 'arm_node'
+    node_name = 'science_node'
     rospy.init_node(node_name, anonymous=False) # only allow one node of this type
-    rospy.loginfo('Initialized "'+node_name+'" node for pub/sub/service functionality')
+    rospy.loginfo('Initialized "' + node_name + '" node for pub/sub/service functionality')
 
     init_serial()
 
-    angle_pub_topic = '/arm_joint_states'
-    rospy.loginfo('Beginning to publish to "'+angle_pub_topic+'" topic')
-    anglePub = rospy.Publisher(angle_pub_topic, JointState, queue_size=10)
-
-    v_bat_topic = '/battery_voltage'
-    rospy.loginfo('Beginning to publish to "'+v_bat_topic+'" topic')
-    vBatPub = rospy.Publisher(v_bat_topic, Float32, queue_size=10)
-
-    feedback_pub_topic = '/arm_feedback'
-    rospy.loginfo('Beginning to publish to "'+feedback_pub_topic+'" topic')
+    feedback_pub_topic = '/science_feedback'
+    rospy.loginfo('Beginning to publish to "' + feedback_pub_topic + '" topic')
     feedbackPub = rospy.Publisher(feedback_pub_topic, String, queue_size=10)
 
-    subscribe_topic = '/arm_command'
-    rospy.loginfo('Beginning to subscribe to "'+subscribe_topic+'" topic')
-    #I could have a topic that listens for JointStateso and constructs the message here...
-    sub = rospy.Subscriber(subscribe_topic, String, subscriber_callback)
-
-    service_name = '/arm_request'
+    service_name = '/science_request'
     rospy.loginfo('Waiting for "'+service_name+'" service request from client')
-    serv = rospy.Service(service_name, ArmRequest, handle_client)
+    serv = rospy.Service(service_name, ScienceRequest, handle_client)
 
     # service requests are implicitly handled but only at the rate the node publishes at
     global ser
     global reqFeedback
-    reqFeedback=''
+    reqFeedback = ''
     global reqInWaiting
-    reqInWaiting=False
+    reqInWaiting = False
     try:
         while not rospy.is_shutdown():
-            #if I try reading from the serial port inside callbacks, bad things happen
-            #instead I send the data elsewhere if required but only read from serial here.
-            #not sure if I need the same precautions when writing but so far it seems ok.
+            # if I try reading from the serial port inside callbacks, bad things happen
+            # instead I send the data elsewhere if required but only read from serial here.
+            # not sure if I need the same precautions when writing but so far it seems ok.
             if ser.in_waiting:
-                data=''
-                feedback=None
+                data = ''
+                feedback = None
                 try:
                     data = ser.readline().decode()
                     feedback = stripFeedback(data)
                 except:
                     rospy.logwarn('trouble reading from serial port')
                 if feedback is not None:
-                    if 'Motor Angles' in feedback:
-                        #rospy.loginfo(feedback)
-                        publish_joint_states(feedback)
-                    elif 'battery voltage' in feedback:
-                        left,voltage = feedback.split('battery voltage: ')
-                        vBatPub.publish(float(voltage))
+                    #rospy.loginfo(feedback)
+                    if reqInWaiting:
+                        reqFeedback += feedback+'\r\n' #pass data to request handler
                     else:
                         #rospy.loginfo(feedback)
-                        if reqInWaiting:
-                            reqFeedback += feedback+'\r\n' #pass data to request handler
-                        else:
-                            if 'WARNING' in feedback:
-                                rospy.logwarn(feedback)
-                            #rospy.loginfo(feedback)
-                            feedbackPub.publish(feedback)
+                        feedbackPub.publish(feedback)
                 else:
-                    rospy.loginfo('got raw data: '+data)
-            rospy.Rate(100).sleep()
+                    rospy.loginfo('got raw data: '+ data)
+            rospy.Rate(50).sleep()
     except rospy.ROSInterruptException:
         pass
 
     def shutdown_hook():
-        rospy.logwarn('This node ('+node_name+') is shutting down')
+        rospy.logwarn('This node (' + node_name + ') is shutting down')
         ser.close() # good practice to close the serial port
         # do I need to clear the serial buffer too?
         time.sleep(1) # give ROS time to deal with the node closing (rosbridge especially)
