+++ robot/rospackages/src/mcu_control/scripts/examples/talker_service.py
