+++ robot/rospackages/src/mcu_control/scripts/ideas
