+++ robot/rospackages/src/mcu_control/scripts/rosbridge_bugfix
